# clients/go/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: ""
    name: ""
  description: ""
  license:
    name: ""
  title: notifier
  version: 0.2.2
servers:
- url: /
paths:
  /add_channel:
    post:
      deprecated: false
      description: |
        Add channel

        Remove notification channel for user
      operationId: handle_add_channel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddChannelBody'
        required: true
      responses:
        "200":
          description: Channel added successfully
      summary: Add channel
      tags:
      - crate
  /get_telegram_chat_id:
    post:
      deprecated: false
      description: "Get the chat ID of a telegram username\n\nFirst call this endpoint,\
        \ then ask the user to send a message to the bot, then the chat id will be\
        \ returned\n"
      operationId: handle_get_telegram_chat_id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTelegramChatIdBody'
        required: true
      responses:
        "200":
          description: Channel added successfully
      summary: Get the chat ID of a telegram username
      tags:
      - crate
  /notify:
    post:
      deprecated: false
      description: |
        Send notification

        send notification to user with given id on all channels registered for that user
      operationId: handle_notify
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNotificationBody'
        required: true
      responses:
        "200":
          description: Notification sent successfully
      summary: Send notification
      tags:
      - crate
  /remove_channel:
    post:
      deprecated: false
      description: |
        Remove channel

        Remove notification channel for user
      operationId: handle_remove_channel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveChannelBody'
        required: true
      responses:
        "200":
          description: Channel removed successfully
      summary: Remove channel
      tags:
      - crate
components:
  schemas:
    RemoveChannelBody:
      example:
        user_id: user_id
        service_id: service_id
      properties:
        user_id:
          type: string
        service_id:
          type: string
      required:
      - service_id
      - user_id
      type: object
    GetTelegramChatIdBody:
      example:
        telegram_username: telegram_username
        user_id: user_id
      properties:
        telegram_username:
          type: string
        user_id:
          type: string
      required:
      - telegram_username
      - user_id
      type: object
    AddChannelBody:
      example:
        user_id: user_id
        service_username: service_username
        service_id: service_id
      properties:
        user_id:
          type: string
        service_username:
          type: string
        service_id:
          type: string
      required:
      - service_id
      - service_username
      - user_id
      type: object
    SendNotificationBody:
      example:
        user_id: user_id
        subject: subject
        message: message
      properties:
        subject:
          type: string
        user_id:
          type: string
        message:
          type: string
      required:
      - message
      - subject
      - user_id
      type: object